tidy up drop shadows
