+++ Content.Stellar.Server/Dropshadow/DropshadowSystem.cs
@@ -1,0 +1,5 @@
+using Content.Stellar.Shared.Dropshadow;
+
+namespace Content.Stellar.Shared.Dropshadow;
+
+public sealed class DropshadowSystem : SharedDropshadowSystem;
