+++ Resources/Prototypes/Entities/Mobs/Species/base.yml
@@ -212,8 +212,9 @@
     - AnomalyHost
   - type: CosmicCenserTarget # Stellar - Cosmic Cult
   - type: Dropshadow # Begin Stellar - Wallening
-    sprite: _ST/Effects/dropshadow.rsi
-    state: shadow-medium
+    sprite:
+      sprite: _ST/Effects/dropshadow.rsi
+      state: shadow-medium
     offset: 0, -0.5 # End Stellar - Wallening
 
 - type: entity
